+++ Bullet.h
@@ -7,7 +7,7 @@ class Bullet : public Entity {
 private:
 	int length;
 public:
-	Bullet(int _pos_x, int _pos_y, int _direction, int _length);
+	Bullet(float _pos_x, float _pos_y, int _direction, int _length);
 	pair<int, int> move_test();
 	bool move();
 	void display();
